+++ src/modules/app/app.vue
@@ -36,8 +36,7 @@
     <authPendingRequestBanner />
 
     <organizationsAdminPendingBanner />
-    <organizationsSuggestedJoinBanner />
-    <organizationsPendingInvitationsNudge />
+    <organizationsLoginNotices />
     <legalCookieBanner />
     <legalFooterSection />
     <v-main class="pb-0" :style="mainStyle">
@@ -65,8 +64,7 @@ import authEmailBanner from '../auth/components/emailBanner.component.vue';
 import authPendingRequestBanner from '../auth/components/pendingRequestBanner.component.vue';
 
 import organizationsAdminPendingBanner from '../organizations/components/organizations.adminPendingBanner.component.vue';
-import organizationsSuggestedJoinBanner from '../organizations/components/organizations.suggestedJoinBanner.component.vue';
-import organizationsPendingInvitationsNudge from '../organizations/components/organizations.pendingInvitationsNudge.component.vue';
+import organizationsLoginNotices from '../organizations/components/organizations.loginNotices.component.vue';
 import legalCookieBanner from '../legal/components/legal.cookieBanner.component.vue';
 import legalFooterSection from '../legal/components/legal.footerSection.component.vue';
 import appErrorBoundary from './components/app.errorBoundary.component.vue';
@@ -84,8 +82,7 @@ export default {
     authPendingRequestBanner,
 
     organizationsAdminPendingBanner,
-    organizationsSuggestedJoinBanner,
-    organizationsPendingInvitationsNudge,
+    organizationsLoginNotices,
     legalCookieBanner,
     legalFooterSection,
     appErrorBoundary,
